requirements.txt
uvloop==0.20.0

fastmcp==1.0
fastapi==0.116.1

aiodns==3.2.0
redis==6.4.0
aioredis==2.0.1
aio-pika==9.4.1

loguru==0.7.2
structlog==24.4.0

serpapi==0.1.5
pydantic-settings==2.10.1

prometheus-client==0.21.1
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0

tenacity==9.0.0

pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

black==24.8.0
isort==5.13.2
mypy==1.11.2
